% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_API.R
\name{search_comex}
\alias{search_comex}
\title{Pesquisa informações no sistema Comex Stat}
\usage{
search_comex(
  ano_inicial = format(Sys.Date(), "\%Y"),
  ano_final = format(Sys.Date(), "\%Y"),
  mes_inicial = 1,
  mes_final = 12,
  detalha_mes = FALSE,
  tipo_op = "exp",
  tipo_ord = "val",
  filtros = character(),
  filtros_esp = list(),
  detalhamentos = character(),
  valor_FOB = TRUE,
  valor_kg = TRUE,
  qtd_est = FALSE
)
}
\arguments{
\item{ano_inicial}{\link{int} Ano inicial da consulta. Ex: 2018.}

\item{ano_final}{\link{int} Ano final da consulta. Ex: 2018.}

\item{mes_inicial}{\link{int} Mês inicial da consulta (1-12).}

\item{mes_final}{\link{int} Mês final da consulta (1-12).}

\item{detalha_mes}{\link{logical} Se TRUE, detalha informações por mês.}

\item{tipo_op}{\link{char} Tipo da operação: \code{'exp'} (exportação) ou \code{'imp'} (importação).}

\item{tipo_ord}{\link{char} Tipo de ordenamento: \code{'val'} (valor) ou \code{'det'} (detalhamento).}

\item{filtros}{\link{character} Vetor com filtros. Ex: \code{c('pais', 'ncm')}.}

\item{filtros_esp}{\link{list} Lista com valores para cada filtro.
Ex: \code{list(c(160,249))}. Consulte \code{\link{Tabelas_Auxiliares}}.}

\item{detalhamentos}{\link{character} Vetor com detalhamentos. Ex: \code{c('pais', 'ncm')}.}

\item{valor_FOB}{\link{logical} Se TRUE, inclui valores FOB.}

\item{valor_kg}{\link{logical} Se TRUE, inclui valores de peso líquido (kg).}

\item{qtd_est}{\link{logical} Se TRUE, inclui quantidade estatística (necessário detalhamento NCM).}
}
\value{
Data frame com os dados retornados pela API, ou \code{NULL} se não houver resultados.
}
\description{
Consulta estatísticas de importação e exportação disponibilizadas pelo sistema
Comex Stat e retorna um data frame com os resultados.
}
\examples{
\dontrun{
pesquisar_comex_stat(
  ano_inicial = 2022, ano_final = 2022,
  mes_inicial = 1, mes_final = 12,
  detalha_mes = TRUE,
  tipo_op = "imp",
  filtros = c("country", "state"),
  filtros_esp = list(c(105, 107), c(26, 13)),
  detalhamentos = c("country", "state", "ncm")
)
}
}
